+++ src/main/java/net/geforcemods/securitycraft/api/Option.java
@@ -174,6 +174,17 @@ public abstract class Option<T> {
 		}
 	}
 
+	public static class IgnoreOwnerOption extends OptionBoolean {
+		public IgnoreOwnerOption(Boolean value) {
+			super("ignoreOwner", value);
+		}
+
+		@Override
+		public String getKey(Block block) {
+			return "option.generic.ignoreOwner";
+		}
+	}
+
 	/**
 	 * A subclass of {@link Option}, set up to handle integers.
 	 */