+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityCageTrap.java
@@ -2,12 +2,14 @@ package net.geforcemods.securitycraft.tileentity;
 
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.DisabledOption;
+import net.geforcemods.securitycraft.api.Option.IgnoreOwnerOption;
 import net.geforcemods.securitycraft.api.Option.OptionBoolean;
 import net.geforcemods.securitycraft.misc.EnumModuleType;
 
 public class TileEntityCageTrap extends TileEntityDisguisable {
 	private OptionBoolean shouldCaptureMobsOption = new OptionBoolean("captureMobs", false);
 	private DisabledOption disabled = new DisabledOption(false);
+	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);
 
 	@Override
 	public EnumModuleType[] acceptedModules() {
@@ -24,10 +26,14 @@ public class TileEntityCageTrap extends TileEntityDisguisable {
 		return disabled.get();
 	}
 
+	public boolean ignoresOwner() {
+		return ignoreOwner.get();
+	}
+
 	@Override
 	public Option<?>[] customOptions() {
 		return new Option[] {
-				shouldCaptureMobsOption, disabled
+				shouldCaptureMobsOption, disabled, ignoreOwner
 		};
 	}
 }