+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityClaymore.java
@@ -2,6 +2,7 @@ package net.geforcemods.securitycraft.tileentity;
 
 import net.geforcemods.securitycraft.api.CustomizableSCTE;
 import net.geforcemods.securitycraft.api.Option;
+import net.geforcemods.securitycraft.api.Option.IgnoreOwnerOption;
 import net.geforcemods.securitycraft.api.Option.OptionInt;
 import net.geforcemods.securitycraft.blocks.mines.BlockClaymore;
 import net.geforcemods.securitycraft.misc.EnumModuleType;
@@ -20,6 +21,7 @@ import net.minecraft.util.math.BlockPos;
 
 public class TileEntityClaymore extends CustomizableSCTE implements ITickable {
 	private OptionInt range = new OptionInt(this::getPos, "range", 5, 1, 10, 1, true);
+	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);
 	private int cooldown = -1;
 
 	@Override
@@ -52,7 +54,7 @@ public class TileEntityClaymore extends CustomizableSCTE implements ITickable {
 			else if (dir == EnumFacing.WEST)
 				area = area.contract(range.get(), -0, -0);
 
-			getWorld().getEntitiesWithinAABB(EntityLivingBase.class, area, e -> !EntityUtils.isInvisible(e) && (!(e instanceof EntityPlayer) || !((EntityPlayer) e).isSpectator()) && !EntityUtils.doesEntityOwn(e, world, pos)).stream().findFirst().ifPresent(entity -> {
+			getWorld().getEntitiesWithinAABB(EntityLivingBase.class, area, e -> !EntityUtils.isInvisible(e) && (!(e instanceof EntityPlayer) || !((EntityPlayer) e).isSpectator()) && !(EntityUtils.doesEntityOwn(e, world, pos) && ignoresOwner())).stream().findFirst().ifPresent(entity -> {
 				cooldown = 20;
 				getWorld().playSound(null, new BlockPos(pos.getX() + 0.5D, pos.getY() + 0.5D, pos.getZ() + 0.5D), SoundEvents.BLOCK_LEVER_CLICK, SoundCategory.BLOCKS, 0.3F, 0.6F);
 			});
@@ -76,12 +78,16 @@ public class TileEntityClaymore extends CustomizableSCTE implements ITickable {
 	@Override
 	public Option<?>[] customOptions() {
 		return new Option[] {
-				range
+				range, ignoreOwner
 		};
 	}
 
 	@Override
 	public EnumModuleType[] acceptedModules() {
 		return new EnumModuleType[0];
+	}
+
+	public boolean ignoresOwner() {
+		return ignoreOwner.get();
 	}
 }