+++ Changelog.md
@@ -4,6 +4,7 @@
 - New: Netherrack Mine and End Stone Mine
 - New: The damage dealt by a Laser Block containing a Harming Module can now be changed with a config setting
 - New: More modded wooden chests can now be converted to a Password-protected Chest
+- New: Ignore owner option for blocks like the Inventory Scanner or Block Change Detector
 - API: Changed IPasswordConvertible#getOriginalBlock to IPasswordConvertible#isValidStateForConversion for finer control over what can be converted
 - API: New methods IModuleInventory#isAllowed and IModuleInventory#isDenied to check whether an entity is listed on an allowlist or denylist module respectively
 - API: Moved and renamed Owner#isOwner methods to IPasswordProtected#isOwnedBy
@@ -17,6 +18,7 @@
 - Fix: Breaking an Inventory Scanner/Laser or one of its field does not break all fields, if the respective config setting is lower than the amount of fields present
 - Fix: Breaking an Inventory Scanner Fields with the inventoryScannerRange config setting set to >=3 does not break and reinstate all fields properly
 - Fix: Players that are on the allowlist of a Security Camera, cannot toggle the camera's redstone output
+- Fix: The Redstone Module in a Laser Block does not respect ownership
 - Fix: Disabled option in laser blocks does not work properly
 - Fix: The Username Logger sometimes cannot be activated by a short redstone signal